Bluetooth: Move DIS implementation to subsys/bluetooth/lib
This moves GATT Device Information Service implementation under
subsys/bluetooth/lib and integrate with Kbuild to avoid build
problems.

Fixes #1166
