+++ samples/bluetooth/peripheral/prj.conf
@@ -9,3 +9,4 @@ CONFIG_BT_DEVICE_NAME="Test peripheral"
 CONFIG_BT_DEVICE_APPEARANCE=833
 CONFIG_BT_BAS=y
 CONFIG_BT_CTS=y
+CONFIG_BT_DIS=y
