+++ samples/bluetooth/peripheral_csc/prj.conf
@@ -5,3 +5,4 @@ CONFIG_BT_PERIPHERAL=y
 CONFIG_BT_DEVICE_NAME="CSC peripheral"
 CONFIG_BT_DEVICE_APPEARANCE=1157
 CONFIG_BT_BAS=y
+CONFIG_BT_DIS=y
