+++ samples/bluetooth/peripheral_dis/prj.conf
@@ -1,3 +1,4 @@
 CONFIG_BT=y
 CONFIG_BT_PERIPHERAL=y
 CONFIG_BT_DEVICE_NAME="DIS peripheral"
+CONFIG_BT_DIS=y
