--- samples/bluetooth/peripheral_dis/src/dis.c
@@ -1,2 +1,0 @@
-/* Workaround build system bug that will put objects in source dir */
-#include "../../gatt/dis.c"
