+++ samples/bluetooth/peripheral_esp/prj.conf
@@ -5,3 +5,4 @@ CONFIG_TINYCRYPT=y
 CONFIG_BT_DEVICE_NAME="ESP peripheral"
 CONFIG_BT_DEVICE_APPEARANCE=768
 CONFIG_BT_BAS=y
+CONFIG_BT_DIS=y
