+++ samples/bluetooth/peripheral_esp/src/Makefile
@@ -1,3 +1,1 @@
-ccflags-y +=-I${ZEPHYR_BASE}/samples/bluetooth
-
-obj-y = main.o dis.o
+obj-y = main.o
