+++ samples/bluetooth/peripheral_esp/src/main.c
@@ -21,8 +21,7 @@
 #include <bluetooth/uuid.h>
 #include <bluetooth/gatt.h>
 #include <bluetooth/bas.h>
-
-#include <gatt/dis.h>
+#include <bluetooth/dis.h>
 
 #define DEVICE_NAME				CONFIG_BT_DEVICE_NAME
 #define DEVICE_NAME_LEN				(sizeof(DEVICE_NAME) - 1)
