+++ samples/bluetooth/peripheral_hids/src/main.c
@@ -20,8 +20,8 @@
 #include <bluetooth/uuid.h>
 #include <bluetooth/gatt.h>
 #include <bluetooth/bas.h>
+#include <bluetooth/dis.h>
 
-#include <gatt/dis.h>
 #include <gatt/hog.h>
 
 #define DEVICE_NAME		CONFIG_BT_DEVICE_NAME
