+++ subsys/bluetooth/lib/Makefile
@@ -1,2 +1,3 @@
 obj-$(CONFIG_BT_BAS) += bas.o
 obj-$(CONFIG_BT_CTS) += cts.o
+obj-$(CONFIG_BT_DIS) += dis.o
